Restructure docs source and output
This PR reorganizes the docs to follow our target top-level structure of

* Overview
* Quickstarts
* Tutorials
* Concepts
* Reference

We have a single toctree for each of those.

It also restructures the source of the docs to mirror the output
hierarchy. Quickstarts go in a `quickstarts` fodler. Tutorials in
`tutorials`, etc.

I've put a couple placeholder files there for now.