Improve cookie consent rendering
Moves the library scripts to the main index page so they are requested
early, and deferred so they don't block rendering. The script to
register the consent is placed in a retry block to attempt to use the
library over a period of a second. This was prompted my flaky test
results from the web compliance portal that indicated some type of race
condition.