Update consent and analytics loading
While the consent and analytics libraries are working in practice, they
continue to fail the automated compliance checks. By moving the loading
and rendering to the body, outside of React, I hope to have the elements
in place prior to the automated test being run. Functionally, there is
little difference for the user, though the banner will display briefly
as the rest of the app renders and then reflows if they browser restores
a scroll position. This seems like an ok tradeoff.