Allow grouped collections
We're developing a STAC based approach to creating subcatalog of
thematically similar collections (think Daymet, GOES, MODIS). Rather
than enumerate them all on our main catalog page, they can be
represented with by a group. The current implementation allows for
manual grouping by keys, with app-specific metadata. It is expected that
in the future this will be derived from the API directly. As such, this
is only an interim solution.