Fix STAC markdown newlines
STAC description field had a \\n for newlines, which caused linebreaks
not to parse. The previous fix that replaced it with a literal <br/> had
the unintended consequence of acting as the first character on a new
line, preventing some markdown syntax from getting parsed. Instead,
replace with an appropriate newline char and the markdown syntax is
restored.