Cancel existing register queries before new request
There was a race condition for setting the hash key used in the mosaic
tile requests based on the most recently returned /register response. In
the case of querystring loading, two requests were fired quickly for the
default case and for the selected case. Either one could have returned
last, and set the hash key