Move mapInfo config to mosaicInfo
The API previously provided default coordinates and zoom for datasets,
but that configuration was tied to the mechanism that provided early
sqlite-db generate mosaics. When this feature was disabled, the mapInfo
was unintentionally disabled. Consolidate this config with mosaic config,
though all of that eventually will migrate back to the API.