Provide a default api root url when none is provided (#326)
Generally this is set in CI or on the dev machine, but for GH Actions
initiated by a non-contrib users which run the unit tests, this value
isn't injected in from the secrets. The error is about parsing a valid
URL for tests, the actual response is mocked so the default value is
not important for this purpose. However, we use staging as a sensible
default for situations when a dev environment is created without the
value set.