Add ImageExporter as a Sidebar panel.
This adds the components for the ImageExporter, the state Slice,
and the requests.

This was mostly copy-pasted from the animation exporter component;
likely there's opportunity to refactor and consolidate shared controls.

Also, added a "next-share" library to add social media
share buttons on the ImageViewer. These can also be added
to the animation exporter.