Add image export sidebar panel (#362)
* drawnBbox -> drawnShape

This stores the drawn shape information in a new
IDrawnShape type on the mapSlice. This allows
storing of both bbox and geometry information,
which will be used in the image export panel.

Note that the atlas.Shape itself cannot be stored
in the mapSlice, because it is not serializable.

* Add ImageExporter as a Sidebar panel.

This adds the components for the ImageExporter, the state Slice,
and the requests.

This was mostly copy-pasted from the animation exporter component;
likely there's opportunity to refactor and consolidate shared controls.

Also, added a "next-share" library to add social media
share buttons on the ImageViewer. These can also be added
to the animation exporter.

* Add image sidebar to context menu

This refactors setShowAnimationPanel -> setShowSidebar and adds
the image panel as an option to the SearchResultHeaderMenu.

* Consolidate animation & image exporter UIs

Co-authored-by: Matt McFarland <mmcfarland@microsoft.com>