Validate bbox before export
The Azure Maps bounding box would occasionally invert the xmin and xmax
values of the bounding box under hard to determine circumstances. This
nearly always happened for geometries that spanned large areas in both
dimensions, but didn't seem to be related to anything incorrect from the
source geometry. As a fail safe, the bbox is coerced into a
xmin,ymin,xmax,ymax order assuming that x and y coordinates remain in
their appropriate slots, but not always the right order.