Add checks for CA Privacy Requirement control
Add a compliance check for ensuring that global privacy opt-out is
respected in the user browser settings or cookies. Planetary Computer
does not use any advertising cookies, but does collect usage information
that is aggregated into 1DS. This means that it is in scope for
compliance with the CPR.

Check if the browser emits a Global Do not Track signal, or that an AMC
(account.ms.com) privacy dashboard cookie is set.