California Privacy Requirement website compliance (#449)
* Add checks for CA Privacy Requirement control

Add a compliance check for ensuring that global privacy opt-out is
respected in the user browser settings or cookies. Planetary Computer
does not use any advertising cookies, but does collect usage information
that is aggregated into 1DS. This means that it is in scope for
compliance with the CPR.

Check if the browser emits a Global Do not Track signal, or that an AMC
(account.ms.com) privacy dashboard cookie is set.

* Add ad choice link to footer

* Upgrade to the minimum viable sphinx version

Some transitive sphinxcontrib-* dependencies are not pinned and now
started requiring sphinx v5+ in order to build our docs. Upgrading to
the most recent v5.* series seems to still build our docs correctly
without additional config changes.

* Lint

* Remove test code